@@ -1,8 +1,6 @@
 import { Restaurant } from './restaurant/restaurant.model';
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
+import { Observable } from 'rxjs';
 import { MEAT_API } from '../app.api';
 import { MenuItem } from '../restaurant-detail/menu-item/menu-item.model';
 import { HttpClient, HttpParams } from '@angular/common/http';
