@@ -1,10 +1,9 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { MEAT_API } from '../../app.api';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { User } from './user.model';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/filter';
+import {tap, filter} from 'rxjs/operators'
 import { Router, NavigationEnd } from '@angular/router';
 
 @Injectable()
@@ -14,7 +13,7 @@ export class LoginService {
     lastUrl: string;
 
     constructor(private http: HttpClient, private router: Router) {
-        this.router.events.filter(e => e instanceof NavigationEnd)
+        this.router.events.pipe(filter(e => e instanceof NavigationEnd))
                           .subscribe( (e: NavigationEnd) => this.lastUrl = e.url);
     }
 
@@ -24,7 +23,7 @@ export class LoginService {
 
     login(email: string, password: string): Observable<User> {
         return this.http.post<User>(`${MEAT_API}/login`, {email: email, password: password})
-                        .do(user => this.user = user);
+                        .pipe(tap(user => this.user = user));
     }
 
     handleLogin(path: string = this.lastUrl) {
