@@ -1,10 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { trigger, state, style, transition, animate } from '@angular/animations';
 import { NotificationService } from '../notification.service';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/timer';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/switchMap';
+import { timer } from 'rxjs';
+import {tap, switchMap} from 'rxjs/operators';
 
 @Component({
   selector: 'mt-snackbar',
@@ -33,11 +31,14 @@ export class SnackbarComponent implements OnInit {
   constructor(private notificationService: NotificationService) { }
 
   ngOnInit() {
-    this.notificationService.notifier.do(message => {
-      this.message = message;
-      this.snackVisibility = 'visible';
-    }).switchMap(message => Observable.timer(3000))
-      .subscribe(timer => this.snackVisibility = 'hidden');
+    this.notificationService.notifier
+    .pipe(
+      tap(message => {
+        this.message = message;
+        this.snackVisibility = 'visible';
+      }),
+     switchMap(_message => timer(3000))
+    ).subscribe(_timer => this.snackVisibility = 'hidden');
   }
 
 }
