@@ -19,7 +19,11 @@ import { OrderSumaryComponent } from './order-sumary/order-sumary.component';
 import { SharedModule } from './shared/shared.module';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import { NotFoundComponent } from './not-found/not-found.component';
-import { LocationStrategy, HashLocationStrategy } from '@angular/common';
+import { LocationStrategy, HashLocationStrategy, registerLocaleData } from '@angular/common';
+import localePt from '@angular/common/locales/pt';
+
+registerLocaleData(localePt, 'pt')
+
 import { LoginComponent } from './security/login/login.component';
 import { UserDetailComponent } from './header/user-detail/user-detail.component';
 import { ApplicationErrorHandler } from './app-error-handler';
@@ -52,7 +56,7 @@ import { ApplicationErrorHandler } from './app-error-handler';
   providers: [
     {provide: LocationStrategy, useClass: HashLocationStrategy},
     {provide: ErrorHandler, useClass: ApplicationErrorHandler},
-    {provide: LOCALE_ID, useValue: 'pt-BR'}],
+    {provide: LOCALE_ID, useValue: 'pt'}],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
